Fix: 마이페이지 버그 해결
