+++ backend/src/main/java/project/backend/domain/chat/chatroom/api/ChatRoomController.java
@@ -40,81 +40,81 @@ import project.backend.global.exception.ex.AuthException;
 @RequestMapping("/chat-rooms")
 public class ChatRoomController {
 
-	private final ChatRoomService chatRoomService;
+    private final ChatRoomService chatRoomService;
 
-	@PostMapping
-	@ResponseBody
-	@ResponseStatus(HttpStatus.CREATED)
-	public ChatRoomSimpleResponse createChatRoom(@Valid @RequestBody ChatRoomRequest request,
-		@AuthenticationPrincipal MemberDetails memberDetails) {
-		Long ownerId = memberDetails.getId();
+    @PostMapping
+    @ResponseBody
+    @ResponseStatus(HttpStatus.CREATED)
+    public ChatRoomSimpleResponse createChatRoom(@Valid @RequestBody ChatRoomRequest request,
+                                                 @AuthenticationPrincipal MemberDetails memberDetails) {
+        Long ownerId = memberDetails.getId();
 
-		log.info("채팅방생성");
-		return chatRoomService.createChatRoom(request, ownerId);
-	}
+        log.info("채팅방생성");
+        return chatRoomService.createChatRoom(request, ownerId);
+    }
 
-	@GetMapping("/invite/{roomId}")
-	@ResponseBody
-	public InviteCodeResponse getInviteUrl(@PathVariable Long roomId,
-		@AuthenticationPrincipal MemberDetails memberDetails
-	) {
+    @GetMapping("/invite/{roomId}")
+    @ResponseBody
+    public InviteCodeResponse getInviteUrl(@PathVariable Long roomId,
+                                           @AuthenticationPrincipal MemberDetails memberDetails
+    ) {
 
-		if (memberDetails == null) {
-			throw new AuthException(AuthErrorCode.UNAUTHORIZED_USER);
-		}
+        if (memberDetails == null) {
+            throw new AuthException(AuthErrorCode.UNAUTHORIZED_USER);
+        }
 
-		Long memberId = memberDetails.getId();
-		boolean isParticipant = chatRoomService.isParticipant(roomId, memberId);
+        Long memberId = memberDetails.getId();
+        boolean isParticipant = chatRoomService.isParticipant(roomId, memberId);
 
-		if (!isParticipant) {
-			throw new AuthException(AuthErrorCode.FORBIDDEN_PARTICIPANT);
-		}
+        if (!isParticipant) {
+            throw new AuthException(AuthErrorCode.FORBIDDEN_PARTICIPANT);
+        }
 
-		String inviteCode = chatRoomService.getInviteCode(roomId);
+        String inviteCode = chatRoomService.getInviteCode(roomId);
 
-		return new InviteCodeResponse(inviteCode);
-	}
-
-
-	@PostMapping("/join")
-	@ResponseStatus(HttpStatus.OK)
-	public InviteJoinResponse joinChatRoom(@RequestBody InviteJoinRequest request,
-		@AuthenticationPrincipal MemberDetails memberDetails
-	) {
-		if (memberDetails == null) {
-			throw new AuthException(AuthErrorCode.UNAUTHORIZED_USER);
-		}
-
-		Long roomId = chatRoomService.joinChatRoom(request.getInviteCode(), memberDetails.getId());
-		return new InviteJoinResponse(roomId);
-	}
+        return new InviteCodeResponse(inviteCode);
+    }
 
 
-	@GetMapping("/recent")
-	public RecentChatRoomResponse getRecentRoomId(Principal principal) {
-		Long roomId = chatRoomService.getMostRecentRoomId(principal.getName());
-		return new RecentChatRoomResponse(roomId);
-	}
+    @PostMapping("/join")
+    @ResponseStatus(HttpStatus.OK)
+    public InviteJoinResponse joinChatRoom(@RequestBody InviteJoinRequest request,
+                                           @AuthenticationPrincipal MemberDetails memberDetails
+    ) {
+        if (memberDetails == null) {
+            throw new AuthException(AuthErrorCode.UNAUTHORIZED_USER);
+        }
 
-	@GetMapping
-	public Page<ChatRoomDetailResponse> findAllChatRooms(
-		@AuthenticationPrincipal MemberDetails memberDetails,
-		@PageableDefault(size = 20, sort = "createdAt", direction = Sort.Direction.DESC) Pageable pageable) {
-		if (memberDetails == null) {
-			throw new AuthException(AuthErrorCode.UNAUTHORIZED_USER);
-		}
-		Long memberId = memberDetails.getId();
-		// 채팅방 목록 리스트로 가져오기
-		return chatRoomService.findChatRoomsByParticipantId(memberId, pageable);
-	}
+        Long roomId = chatRoomService.joinChatRoom(request.getInviteCode(), memberDetails.getId());
+        return new InviteJoinResponse(roomId);
+    }
 
-  // 자신이 만든 채팅방 가져오기 -> 주후 인증객체 id로 조회가능 할듯(Authentication)
-  @GetMapping("/mine/{memberId}")
-  public Page<MyChatRoomResponse> findMyAllChatRooms(@PathVariable Long memberId,
-                                                     @PageableDefault(size = 10, sort = "createdAt", direction = Sort.Direction.DESC) Pageable pageable) {
-    log.info("자신이 만든 채팅방 요청: memberId = {}", memberId);
-    return chatRoomService.findAllRoomsByOwnerId(memberId, pageable);
 
-  }
+    @GetMapping("/recent")
+    public RecentChatRoomResponse getRecentRoomId(Principal principal) {
+        Long roomId = chatRoomService.getMostRecentRoomId(principal.getName());
+        return new RecentChatRoomResponse(roomId);
+    }
+
+    @GetMapping
+    public Page<ChatRoomDetailResponse> findAllChatRooms(
+            @AuthenticationPrincipal MemberDetails memberDetails,
+            @PageableDefault(size = 20, sort = "createdAt", direction = Sort.Direction.DESC) Pageable pageable) {
+        if (memberDetails == null) {
+            throw new AuthException(AuthErrorCode.UNAUTHORIZED_USER);
+        }
+        Long memberId = memberDetails.getId();
+        // 채팅방 목록 리스트로 가져오기
+        return chatRoomService.findChatRoomsByParticipantId(memberId, pageable);
+    }
+
+    // 자신이 만든 채팅방 가져오기 -> 주후 인증객체 id로 조회가능 할듯(Authentication)
+    @GetMapping("/mine/{memberId}")
+    public Page<MyChatRoomResponse> findMyAllChatRooms(@PathVariable Long memberId,
+                                                       @PageableDefault(size = 10, sort = "createdAt", direction = Sort.Direction.DESC) Pageable pageable) {
+        log.info("자신이 만든 채팅방 요청: memberId = {}", memberId);
+        return chatRoomService.findAllRoomsByOwnerId(memberId, pageable);
+
+    }
 
 }
