+++ backend/src/main/java/project/backend/domain/chat/chatroom/dao/ChatRoomRepository.java
@@ -12,12 +12,13 @@ import project.backend.domain.chat.chatroom.entity.ChatRoom;
 
 public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
 
-	@EntityGraph(attributePaths = {"participants", "participants.participant"})
-	Page<ChatRoom> findByParticipants_Participant_Id(Long memberId, Pageable pageable);
+    @EntityGraph(attributePaths = {"participants", "participants.participant"})
+    Page<ChatRoom> findByParticipants_Participant_Id(Long memberId, Pageable pageable);
 
-  Optional<ChatRoom> findByInviteCode(String inviteCode);
+    Optional<ChatRoom> findByInviteCode(String inviteCode);
 
-  Page<ChatRoom> findAllRoomsByOwnerId(Long ownerId, Pageable pageable);
+    
+    Page<ChatRoom> findAllRoomsByOwnerId(Long ownerId, Pageable pageable);
 
 }
