+++ backend/src/main/java/project/backend/domain/member/app/MemberService.java
@@ -105,7 +105,7 @@ public class MemberService {
     public MemberResponse getMemberDetails(Authentication auth) {
         MemberDetails loginMember = (MemberDetails) auth.getPrincipal();
         Long memberId = loginMember.getId();
-
+        log.info("memberId = {}", memberId);
         Member member = getMemberById(memberId);
         return MemberMapper.toResponse(member);
     }
