+++ frontend/src/profile-page.module.css
@@ -229,4 +229,5 @@ body {
   justify-content: center;
   height: 160px;
   color: #6b7280;
+  font-size: 20px;
 }
